Use const assertion for flat config
When using the flat config in a typechecked ESLint file (with
`exactOptionalPropertyTypes: true`), a type error is reported
because the generated type for `flat/recommended` is a union, where
one element in the union contains `files?: undefined`. `files` can be
omitted, but it cannot be explicitly `undefined``.

This uses a const assertion so that, in the inferred type, `files` is
correctly omitted.